Created a new function "getForecastForWeek"
This function is to get the latitude and longitude of every city to get the weekly forecast. Also this function in getting a response inside the showTemperature function.
I included a new url from the One Call API , and the function (showWeatherForecast) that was at the end, I relocated inside this new function.